requirements.txt
torch==1.7.1
torchvision==0.8.2
opencv-python==3.4.2
pillow==8.1.0
numpy==1.10.2